+++ lib/src/enums/torch_state.dart
@@ -1,5 +1,10 @@
 /// The state of the flashlight.
 enum TorchState {
+  /// The flashlight turns on automatically in low light conditions.
+  ///
+  /// This is currently only supported on iOS and MacOS.
+  auto(2),
+
   /// The flashlight is off.
   off(0),
 
@@ -7,18 +12,20 @@ enum TorchState {
   on(1),
 
   /// The flashlight is unavailable.
-  unavailable(2);
+  unavailable(-1);
 
   const TorchState(this.rawValue);
 
   factory TorchState.fromRawValue(int value) {
     switch (value) {
+      case -1:
+        return TorchState.unavailable;
       case 0:
         return TorchState.off;
       case 1:
         return TorchState.on;
       case 2:
-        return TorchState.unavailable;
+        return TorchState.auto;
       default:
         throw ArgumentError.value(value, 'value', 'Invalid raw value.');
     }