+++ lib/src/mobile_scanner.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:mobile_scanner/src/mobile_scanner_controller.dart';
 import 'package:mobile_scanner/src/mobile_scanner_exception.dart';
 import 'package:mobile_scanner/src/mobile_scanner_platform_interface.dart';
+import 'package:mobile_scanner/src/objects/barcode_capture.dart';
 import 'package:mobile_scanner/src/objects/mobile_scanner_state.dart';
 import 'package:mobile_scanner/src/scan_window_calculation.dart';
 
@@ -18,7 +19,8 @@ typedef MobileScannerErrorBuilder = Widget Function(
 class MobileScanner extends StatefulWidget {
   /// Create a new [MobileScanner] using the provided [controller].
   const MobileScanner({
-    required this.controller,
+    this.controller,
+    this.onDetect,
     this.fit = BoxFit.cover,
     this.errorBuilder,
     this.overlayBuilder,
@@ -29,7 +31,11 @@ class MobileScanner extends StatefulWidget {
   });
 
   /// The controller for the camera preview.
-  final MobileScannerController controller;
+  final MobileScannerController? controller;
+
+  /// The function that signals when new codes were detected by the [controller].
+  /// If null, use the controller.barcodes stream directly to capture barcodes.
+  final void Function(BarcodeCapture barcodes)? onDetect;
 
   /// The error builder for the camera preview.
   ///
@@ -112,7 +118,10 @@ class MobileScanner extends StatefulWidget {
   State<MobileScanner> createState() => _MobileScannerState();
 }
 
-class _MobileScannerState extends State<MobileScanner> {
+class _MobileScannerState extends State<MobileScanner>
+    with WidgetsBindingObserver {
+  late final controller = widget.controller ?? MobileScannerController();
+
   /// The current scan window.
   Rect? scanWindow;
 
@@ -139,7 +148,7 @@ class _MobileScannerState extends State<MobileScanner> {
     if (scanWindow == null) {
       scanWindow = newScanWindow;
 
-      unawaited(widget.controller.updateScanWindow(scanWindow));
+      unawaited(controller.updateScanWindow(scanWindow));
 
       return;
     }
@@ -154,7 +163,7 @@ class _MobileScannerState extends State<MobileScanner> {
     if (widget.scanWindowUpdateThreshold == 0.0) {
       scanWindow = newScanWindow;
 
-      unawaited(widget.controller.updateScanWindow(scanWindow));
+      unawaited(controller.updateScanWindow(scanWindow));
 
       return;
     }
@@ -167,14 +176,14 @@ class _MobileScannerState extends State<MobileScanner> {
         dy >= widget.scanWindowUpdateThreshold) {
       scanWindow = newScanWindow;
 
-      unawaited(widget.controller.updateScanWindow(scanWindow));
+      unawaited(controller.updateScanWindow(scanWindow));
     }
   }
 
   @override
   Widget build(BuildContext context) {
     return ValueListenableBuilder<MobileScannerState>(
-      valueListenable: widget.controller,
+      valueListenable: controller,
       builder: (BuildContext context, MobileScannerState value, Widget? child) {
         if (!value.isInitialized) {
           const Widget defaultPlaceholder = ColoredBox(color: Colors.black);
@@ -234,10 +243,62 @@ class _MobileScannerState extends State<MobileScanner> {
     );
   }
 
+  StreamSubscription? _subscription;
+
+  @override
+  void initState() {
+    if (widget.onDetect != null) {
+      WidgetsBinding.instance.addObserver(this);
+      _subscription = controller.barcodes.listen(widget.onDetect);
+    }
+    if (controller.autoStart) {
+      controller.start();
+    }
+    super.initState();
+  }
+
   @override
   void dispose() {
     super.dispose();
+
+    if (_subscription != null) {
+      _subscription!.cancel();
+      _subscription = null;
+    }
+
+    if (controller.autoStart) {
+      controller.stop();
+    }
     // When this widget is unmounted, reset the scan window.
-    unawaited(widget.controller.updateScanWindow(null));
+    unawaited(controller.updateScanWindow(null));
+
+    // Dispose default controller if not provided by user
+    if (widget.controller == null) {
+      controller.dispose();
+      WidgetsBinding.instance.removeObserver(this);
+    }
+  }
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    if (widget.controller != null) return;
+    if (!controller.value.isInitialized) {
+      return;
+    }
+
+    switch (state) {
+      case AppLifecycleState.detached:
+      case AppLifecycleState.hidden:
+      case AppLifecycleState.paused:
+        return;
+      case AppLifecycleState.resumed:
+        _subscription = controller.barcodes.listen(widget.onDetect);
+
+        unawaited(controller.start());
+      case AppLifecycleState.inactive:
+        unawaited(_subscription?.cancel());
+        _subscription = null;
+        unawaited(controller.stop());
+    }
   }
 }